break out AuthDB and BibDB into separate class files, hopefully set up for unit testing
