+++ browse-handler/java/org/vufind/solr/handler/BrowseRequestHandler.java
@@ -272,238 +272,6 @@ class HeadingsDB
 }
 
 
-
-/**
- *
- * Interface to the Solr Authority DB
- *
- */
-class AuthDB
-{
-    static int MAX_PREFERRED_HEADINGS = 1000;
-
-    private SolrIndexSearcher searcher;
-    private String preferredHeadingField;
-    private String useInsteadHeadingField;
-    private String seeAlsoHeadingField;
-    private String scopeNoteField;
-
-    public AuthDB(SolrIndexSearcher authSearcher,
-                  String preferredField,
-                  String useInsteadField,
-                  String seeAlsoField,
-                  String noteField)
-    throws Exception
-    {
-        searcher = authSearcher;
-        preferredHeadingField = preferredField;
-        useInsteadHeadingField = useInsteadField;
-        seeAlsoHeadingField = seeAlsoField;
-        scopeNoteField = noteField;
-    }
-
-
-    private List<String> docValues(Document doc, String field)
-    {
-        String values[] = doc.getValues(field);
-
-        if (values == null) {
-            values = new String[] {};
-        }
-
-        return Arrays.asList(values);
-    }
-
-
-    public Document getAuthorityRecord(String heading)
-    throws Exception
-    {
-        TopDocs results = (searcher.search(new TermQuery(new Term(preferredHeadingField,
-                                           heading)),
-                                           1));
-
-        if (results.totalHits > 0) {
-            return searcher.getIndexReader().document(results.scoreDocs[0].doc);
-        } else {
-            return null;
-        }
-    }
-
-
-    public List<Document> getPreferredHeadings(String heading)
-    throws Exception
-    {
-        TopDocs results = (searcher.search(new TermQuery(new Term(useInsteadHeadingField,
-                                           heading)),
-                                           MAX_PREFERRED_HEADINGS));
-
-        List<Document> result = new ArrayList<> ();
-
-        for (int i = 0; i < results.totalHits; i++) {
-            result.add(searcher.getIndexReader().document(results.scoreDocs[i].doc));
-        }
-
-        return result;
-    }
-
-
-    public Map<String, List<String>> getFields(String heading)
-    throws Exception
-    {
-        Document authInfo = getAuthorityRecord(heading);
-
-        Map<String, List<String>> itemValues = new HashMap<> ();
-
-        itemValues.put("seeAlso", new ArrayList<String>());
-        itemValues.put("useInstead", new ArrayList<String>());
-        itemValues.put("note", new ArrayList<String>());
-
-        if (authInfo != null) {
-            for (String value : docValues(authInfo, seeAlsoHeadingField)) {
-                itemValues.get("seeAlso").add(value);
-            }
-
-            for (String value : docValues(authInfo, scopeNoteField)) {
-                itemValues.get("note").add(value);
-            }
-        } else {
-            List<Document> preferredHeadings =
-                getPreferredHeadings(heading);
-
-            for (Document doc : preferredHeadings) {
-                for (String value : docValues(doc, preferredHeadingField)) {
-                    itemValues.get("useInstead").add(value);
-                }
-            }
-        }
-
-        return itemValues;
-    }
-}
-
-
-
-/**
- *
- * Interface to the Solr biblio db
- *
- */
-class BibDB
-{
-    private IndexSearcher db;
-    private String field;
-
-    public BibDB(IndexSearcher searcher, String field) throws Exception
-    {
-        this.db = searcher;
-        this.field = field;
-    }
-
-    /**
-     * Returns the number of bib records that match an authority heading.
-     * 
-     * @param heading
-     * @return	number of matching bib records
-     * @throws Exception
-     */
-    public int recordCount(String heading)
-    throws IOException
-    {
-        TermQuery q = new TermQuery(new Term(field, heading));
-
-        Log.info("Searching '" + field + "' for '" + "'" + heading + "'");
-
-        TotalHitCountCollector counter = new TotalHitCountCollector();
-        db.search(q, counter);
-
-        Log.info("Hits: " + counter.getTotalHits());
-
-        return counter.getTotalHits();
-    }
-
-
-    /**
-     *
-     * Function to retrieve the doc ids when there is a building limit
-     * This retrieves the doc ids for an individual heading
-     *
-     * Need to add a filter query to limit the results from Solr
-     *
-     * Includes functionality to retrieve additional info
-     * like titles for call numbers, possibly ISBNs
-     *
-     * @param heading        string of the heading to use for finding matching
-     * @param fields         docs colon-separated string of Solr fields
-     *                       to return for use in the browse display
-     * @param maxBibListSize maximum numbers of records to check for fields
-     * @return         return a map of Solr ids and extra bib info
-     */
-    public Map<String, List<Collection<String>>> matchingIDs(String heading, 
-                                                             String fields,
-                                                             int maxBibListSize)
-    throws Exception
-    {
-        TermQuery q = new TermQuery(new Term(field, heading));
-
-        // bibinfo values are List<Collection> because some extra fields
-        // may be multi-valued.
-        // Note: it may be time for bibinfo to become a class...
-        final Map<String, List<Collection<String>>> bibinfo = new HashMap<> ();
-        // Forcing "ids" into list of bib fields is a transition to requiring
-        // that "ids" be listed explicitly in the extras string
-        final String[] bibFieldList = fields.split(":");
-        for (String bibField : bibFieldList) {
-            bibinfo.put(bibField, new ArrayList<Collection<String>> ());
-        }
-
-        db.search(q, new SimpleCollector() {
-            private LeafReaderContext context;
-
-            public void setScorer(Scorer scorer) {
-            }
-
-            public boolean acceptsDocsOutOfOrder() {
-                return true;
-            }
-
-            public boolean needsScores() {
-                return false;
-            }
-
-            public void doSetNextReader(LeafReaderContext context) {
-                this.context = context;
-            }
-
-
-            public void collect(int docnum) {
-                int docid = docnum + context.docBase;
-                try {
-                    Document doc = db.getIndexReader().document(docid);
-
-                    for (String bibField : bibFieldList) {
-                        String[] vals = doc.getValues(bibField);
-                        if (vals.length > 0) {
-                            Collection<String> valSet = new LinkedHashSet<> ();
-                            for (String val : vals) {
-                                valSet.add(val);
-                            }
-                            bibinfo.get(bibField).add(valSet);
-                        }
-                    }
-                } catch (org.apache.lucene.index.CorruptIndexException e) {
-                    Log.info("CORRUPT INDEX EXCEPTION.  EEK! - " + e);
-                } catch (Exception e) {
-                    Log.info("Exception thrown: " + e);
-                }
-
-            }
-        });
-
-        return bibinfo;
-    }
-}
-
-
 /**
  * Class that performs the alphabetical browse of an index and produces a {@code BrowseList} object. 
  *
